// src/types.ts
export enum UserRole {
  MASTER = 'MASTER',
  OWNER = 'OWNER',
  ADMIN_OWNER = 'ADMIN_OWNER',
  ADMIN_DIVISI = 'ADMIN_DIVISI',
  MANAGER = 'MANAGER',
  KASIR = 'KASIR',
  WAREHOUSE = 'WAREHOUSE',
  STAFF = 'STAFF',
}
// ...
export enum BusinessModule {
  POS = 'POS',
  INVENTORY = 'INVENTORY',
  TRANSFER = 'TRANSFER',
  DAMAGED_GOODS = 'DAMAGED_GOODS',
  FINANCE = 'FINANCE',
  ATTENDANCE = 'ATTENDANCE',
  REPORTS = 'REPORTS',
  STAN_OUTLET = 'STAN_OUTLET',
}
// ...
export enum LedgerType {
  PEMASUKAN = 'PEMASUKAN',
  PENGELUARAN = 'PENGELUARAN',
}